requirements.txt
Flask==3.1.2
python-dotenv==1.0.1
# Core LangChain
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5
# Google Integration (Pinned to avoid conflicts)
langchain-google-genai==2.0.1
# Vector Store
langchain-pinecone==0.2.0
pinecone-client==5.0.1
# Graph
langgraph==0.2.45
# Utilities
pymupdf==1.24.11
boto3==1.35.54
python-jose==3.3.0
gunicorn==23.0.0
Pillow==11.0.0
# Reranking
sentence-transformers==3.3.1
einops==0.8.0